Changed the tests so that they use the whole Xacro processing pipeline, not just parts of it.
In order to achieve this, I split the main() function into parsing command-line
arguments part, and then the part responsible for processing.

Also added a file mocking capability to the test suite - it is needed to
fake the input file to the main_process() function.

This simplifies several existing tests, especially these using files on disk.