Solution for #63.
Added an optional "preprocess" boolean attribute to <xacro:if> and <xacro:unless>,
which makes it possible to discard parts of the Xacro file based on <arg>
values without these parts being even touched by Xacro.

This means no $(find ...) expressions in includes will be evaluated, as well
as <xacro:property> tags will also be ignored.

This change works well with the --deps and --include options, so if the
user gives some command-line arguments, he only receives the dependencies
and includes that are really left there after executing preprocess branching.