Fix resolving of macros declared and called in/from a namespace (#297)
Given two macros, declared in a namespace-included file. From outside the namespace, obviously, they need to be called with the namespace prefix.
However, when calling one macro from the other (within the namespace), the prefix shouldn't be required anymore (it's not even known). To fix this issue, the "active" `macros` dictionary needs to be updated during macro resolution.